Made neccessary changes
