+++ glide.lock
@@ -1,39 +1,41 @@
-hash: 50a37de1b666ad3c26385b36d926cee932537fddf8b1fd907f7345ff8eb1a3fb
-updated: 2016-12-28T23:33:37.725385054+05:30
+hash: 44c97b2d3200b1e0300fe1e39dd27969e9eafdd9bd898a2e671759e6d45034c7
+updated: 2016-12-30T22:35:14.940028561+05:30
 imports:
 - name: github.com/jinzhu/gorm
   version: 5174cc5c242a728b435ea2be8a2f7f998e15429b
   subpackages:
+  - dialects/postgres
   - dialects/sqlite
 - name: github.com/jinzhu/inflection
   version: 74387dc39a75e970e7a3ae6a3386b5bd2e5c5cff
 - name: github.com/labstack/echo
   version: 04e6901d05b58ddbae579a6252a3d1bb3f334adc
+  subpackages:
+  - context
+  - engine
+  - log
 - name: github.com/labstack/gommon
   version: b2765095a572012ea4828c9596fefd824c2aee72
   subpackages:
   - color
   - log
+- name: github.com/lib/pq
+  version: 5bf161122cd640c2a5a2c1d7fa49ea9befff31dd
+  subpackages:
+  - hstore
+  - oid
 - name: github.com/mattn/go-colorable
   version: d228849504861217f796da67fae4f6e347643f15
 - name: github.com/mattn/go-isatty
   version: 30a891c33c7cde7b02a981314b4228ec99380cca
 - name: github.com/mattn/go-sqlite3
   version: 2d44decb4941c9cdf72c22297b7890faf7da9bcb
-- name: github.com/tylerb/graceful
-  version: 4df1190835320af7076dfcf27b3d071fd3612caf
 - name: github.com/valyala/fasttemplate
   version: 3b874956e03f1636d171bda64b130f9135f42cff
-- name: golang.org/x/crypto
-  version: f6b343c37ca80bfa8ea539da67a0b621f84fab1d
-  subpackages:
-  - acme
-  - acme/autocert
 - name: golang.org/x/net
   version: 45e771701b814666a7eb299e6c7a57d0b1799e91
   subpackages:
   - context
-  - context/ctxhttp
 - name: golang.org/x/sys
   version: d75a52659825e75fff6158388dddc6a5b04f9ba5
   subpackages:
