+++ glide.yaml
@@ -4,3 +4,4 @@ import:
   version: "2.1"
 - package: github.com/jinzhu/gorm
   version: v1.0
+- package: github.com/lib/pq
